LPS-66799 Apply the new approach to DDM
