+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-service/build.gradle
@@ -8,7 +8,6 @@ dependencies {
 	provided group: "com.liferay", name: "com.liferay.dynamic.data.mapping.data.provider", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.dynamic.data.mapping.io", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.dynamic.data.mapping.validator", version: "2.0.0"
-	provided group: "com.liferay", name: "com.liferay.exportimport.api", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.mobile.device.rules.api", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.osgi.service.tracker.collections", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.osgi.util", version: "3.0.0"
@@ -28,4 +27,5 @@ dependencies {
 	provided group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:forms-and-workflow:dynamic-data-mapping:dynamic-data-mapping-api")
+	provided project(":apps:web-experience:export-import:export-import-api")
 }
