+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-web/build.gradle
@@ -5,8 +5,6 @@ dependencies {
 	provided group: "com.liferay", name: "com.liferay.application.list.api", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.dynamic.data.mapping.io", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.dynamic.data.mapping.validator", version: "2.1.0"
-	provided group: "com.liferay", name: "com.liferay.exportimport.api", version: "2.0.0"
-	provided group: "com.liferay", name: "com.liferay.exportimport.service", version: "3.0.0"
 	provided group: "com.liferay", name: "com.liferay.frontend.taglib", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.petra.content", version: "1.0.0"
 	provided group: "com.liferay", name: "com.liferay.portal.configuration.metatype", version: "2.0.0"
@@ -21,6 +19,8 @@ dependencies {
 	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	provided project(":apps:forms-and-workflow:dynamic-data-mapping:dynamic-data-mapping-api")
 	provided project(":apps:forms-and-workflow:dynamic-data-mapping:dynamic-data-mapping-service")
+    provided project(":apps:web-experience:export-import:export-import-api")
+    provided project(":apps:web-experience:export-import:export-import-service")
 }
 
 mergeLang {
