+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-web/src/main/java/com/liferay/dynamic/data/mapping/web/internal/exportimport/content/processor/DDMTemplateDocumentLibraryExportImportContentProcessor.java
@@ -14,9 +14,22 @@
 
 package com.liferay.dynamic.data.mapping.web.internal.exportimport.content.processor;
 
+import com.liferay.dynamic.data.mapping.model.DDMTemplate;
+import com.liferay.exportimport.content.processor.ExportImportContentProcessor;
+import com.liferay.exportimport.content.processor.base.BaseDocumentLibraryExportImportContentProcessor;
+
+import org.osgi.service.component.annotations.Component;
+
 /**
- * @author Gergely Mathe
+ * @author Mate Thurzo
  */
-@Deprecated
-public class DDMTemplateExportImportContentProcessor {
+@Component(
+	property = {
+		"content.processor.order=1",
+		"model.class.name=com.liferay.dynamic.data.mapping.model.DDMTemplate"
+	},
+	service = {ExportImportContentProcessor.class}
+)
+public class DDMTemplateDocumentLibraryExportImportContentProcessor
+	extends BaseDocumentLibraryExportImportContentProcessor<DDMTemplate> {
 }
